139 moved scripts to final version
