--- Assets/Sandbox/Vlad/Scripts/PlayerDeathBehaviour.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: df60fdb5c83c3cb4dafec139ae84d434
-timeCreated: 1515440010
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
